+++ oap-server/analyzer/agent-analyzer/src/main/java/org/apache/skywalking/oap/server/analyzer/provider/trace/parser/SegmentParserServiceImpl.java
@@ -39,4 +39,5 @@ public class SegmentParserServiceImpl implements ISegmentParserService {
         final TraceAnalyzer traceAnalyzer = new TraceAnalyzer(moduleManager, listenerManager, config);
         traceAnalyzer.doAnalysis(segment);
     }
+
 }